Add MobilNet Model for Machine Learning
