+++ MLendPoint/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Example Node JS with Routes",
   "main": "app.js",
-    "scripts": {
+  "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -19,6 +19,8 @@
   "homepage": "https://github.com/jchoque87/Prog101Routers#readme",
   "devDependencies": {
     "@tensorflow-models/coco-ssd": "^2.2.2",
+    "@tensorflow-models/mobilenet": "^2.1.0",
+    "@tensorflow/tfjs": "^3.8.0",
     "@tensorflow/tfjs-node": "^3.8.0",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
